Etapa5/dash_app/requirements.txt
dash
pandas
plotly
psycopg2-binary